Moved all topic handling to Tower PA system
