+++ bots/drifter.py
@@ -97,35 +97,3 @@ async def chat(interaction: discord.Interaction, prompt: str, temperature: float
 @drifter.bot.tree.command(name="drifter_reset", description="Reset the /drifter_chat AI's memory in case he gets too far gone")
 async def drifter_reset(interaction: discord.Interaction):
     await drifter.text.reset(interaction)
-
-#* Shows the list of random topics to be used daily or with the /generate_conversation command
-@drifter.bot.tree.command(name="drifter_topics", description="View the saved topics that the bots can chat over!")
-async def topics(interaction: discord.Interaction):
-    topics = json.load(open('data/topics.json'))
-    response = ""
-    for _, (key, value) in enumerate(topics.items()):
-        response += f'**{key}:**\n'
-        for v in value["topics"]:
-            response += f'{v}\n'
-        response += '\n'
-    await interaction.response.send_message(f'Heh, listen to this brother. Those \'Disciples\' you killed are wanting to talk about these topics: \n\n{response}', ephemeral=True)
-
-#* Add a topic to the topic list
-@drifter.bot.tree.command(name="drifter_add_topic", description="Add a topic that can be used for the daily conversation!")
-@app_commands.describe(topic="What topic should be added to the list?")
-async def drifter_add_topic(interaction: discord.Interaction, topic: str=None):
-    if topic != None:
-        topics = json.load(open('data/topics.json'))
-        if topic not in topics['misc']["topics"]:
-            topics['misc']["topics"][topic] = { "chosen": False,
-                                                "req_membs": ["all"]}
-            with open('data/topics.json', 'w') as f:
-                log = open('log.txt', 'a')
-                f.write(json.dumps(topics, indent=4))
-                log.write(f'Added a new topic to the list: {topic}\n\n')
-                log.close()
-                await interaction.response.send_message(f'Ooooh {interaction.user.global_name}! **{topic}** sounds like a fun thing to discuss!')
-        else:
-            await interaction.response.send_message(f'{interaction.user.global_name}, did that Primeval drain your Light? (Already in list)')
-    else:
-        await interaction.response.send_message(f'I\'ve been more entertained being cooped up down here than hearing that, {interaction.user.global_name}. (Must input something)')
